increment numServed for customers not entering the store

+++ src/simModel/Output.java
@@ -5,8 +5,8 @@ class Output
 	SMMarket model;
 	
 	/* SSOVs */
-	protected int numServed;
-	protected int numDissatisfied;
+	protected double numServed;
+	protected double numDissatisfied;
 	protected double[] halfHourPercentDissatisfied = new double[18];
 	protected double overallPercentDissatisfied;
